Rename 'updater.rs' to 'index_updater.rs'
