+++ bee-protocol/src/worker/index_updater.rs
